requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
feedparser==6.0.11
APScheduler==3.10.4
PyYAML==6.0.2
requests==2.32.3
